pyproject.toml
[project]
name = "Telegram-Gpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio==4.8.0",
    "asyncio-atexit==1.0.1",
    "beautifulsoup4==4.13.3",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "deeperseek==0.1.4",
    "deprecated==1.2.18",
    "exceptiongroup==1.2.2",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "inscriptis==2.5.3",
    "lxml==5.3.1",
    "mss==10.0.0",
    "python-dotenv==1.0.1",
    "python-telegram-bot==21.10",
    "pyvirtualdisplay==3.0",
    "requests==2.32.3",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "websockets==15.0",
    "wrapt==1.17.2",
    "zendriver==0.5.1",
]
